Arreglar RCView de Vuelos con nueva estructura
