+++ app/src/main/java/com/example/parcial/adapters/TripAdapter.kt
@@ -5,14 +5,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.parcial.R
-import com.example.parcial.activities.SearchResultsActivity
 import com.example.parcial.holders.TripHolder
-import com.example.parcial.listeners.OnViewItemClickedListener
 import com.example.parcial.entities.Trip
+import com.example.parcial.listeners.OnViewItemClickedListener
 
 class TripAdapter(
     private val trips: MutableList<Trip>,
-    //private val onItemClick: OnViewItemClickedListener<Trip>
+    private val onItemClick: OnViewItemClickedListener<Trip>
 
 ): RecyclerView.Adapter<TripHolder>() {
 
@@ -39,11 +38,10 @@ class TripAdapter(
         holder.setAirlineLogo(trip.airlineLogo)
         holder.setTravelClass(trip.travelClass!!)
 
-        /*
         holder.getDetailButton().setOnClickListener {
             onItemClick.onViewItemDetail(trip)
         }
-        */
+
 
     }
 }
